add more code to test

+++ src/example/example1/example1.cc
@@ -3,11 +3,14 @@
 #include "os/ObjectStore.h"
 #include "global/global_context.h"
 #include "perfglue/heap_profiler.h"
+#include "common/errno.h"
 #include <unistd.h>
 #include <iostream>
 #include <map>
 #include <memory>
 #include <string>
+#include <vector>
+#include <semaphore.h>
 
 #define dout_context g_ceph_context
 #define dout_subsys ceph_subsys_
@@ -33,6 +36,368 @@ static void usage()
 {
 	std::cout << "usage: example1 -i <ID> [flags]\n";
 	generic_usage(true);
+}
+
+#define COLLECTION_COUNT 10
+
+// struct to store ObjectStore collection info
+struct MyCollection {
+	spg_t pg;
+	coll_t cid;
+	ObjectStore::CollectionHandle ch;
+	std::unique_ptr<std::mutex> lock;
+
+	static constexpr int64_t MIN_POOL_ID = 0x0000ffffffffffff;
+
+	MyCollection(const spg_t& pg, ObjectStore::CollectionHandle _ch)
+		: pg(pg), cid(pg), ch(_ch),
+	lock(new std::mutex) {                                                  
+	}  
+};
+
+struct Object {                                                                 
+  ghobject_t oid;
+  MyCollection& coll;
+
+  Object(const char* name, MyCollection& coll)
+    : oid(hobject_t(name, "", CEPH_NOSNAP, coll.pg.ps(), coll.pg.pool(), "")),
+      coll(coll) {}
+};
+
+struct Result {
+	sem_t sem_;
+	int result_;
+
+	Result() {
+		sem_init(&sem_, 0, 0);
+		result_ = -1;
+	}
+
+	~Result() {
+		sem_destroy(&sem_);
+	}
+
+	void wait() {
+		while(sem_wait(&sem_) == -1 && errno == EINTR)
+			;
+	}
+
+	void signal(int r) {
+		result_ = r;
+		sem_post(&sem_);
+	}
+
+	int result() const {
+		return result_;
+	}
+};
+
+
+std::vector<MyCollection> collections;
+
+int destroy_collections(
+	std::unique_ptr<ObjectStore>& os,
+	std::vector<MyCollection>& collections);
+
+int init_collections(std::unique_ptr<ObjectStore>& os,                          
+                      uint64_t pool,                                            
+                      std::vector<MyCollection>& collections,
+                      uint64_t count)
+{
+	ceph_assert(count > 0);
+	collections.reserve(count);
+
+	const int split_bits = cbits(count - 1);
+
+	{
+		// propagate Superblock object to ensure proper functioning of tools that
+		// need it. E.g. ceph-objectstore-tool
+		coll_t cid(coll_t::meta());
+		bool exists = os->collection_exists(cid);
+		if (!exists) {
+			auto ch = os->create_new_collection(cid);
+
+			OSDSuperblock superblock;
+			bufferlist bl;
+			encode(superblock, bl);
+
+			ObjectStore::Transaction t;
+			t.create_collection(cid, split_bits);
+			t.write(cid, OSD_SUPERBLOCK_GOBJECT, 0, bl.length(), bl);
+			int r = os->queue_transaction(ch, std::move(t));
+
+			if (r < 0) {
+				std::cerr << "Failure to write OSD superblock: " << cpp_strerror(-r) << std::endl;
+				return r;
+			}
+		}
+	}
+
+	for (uint32_t i = 0; i < count; i++) {
+		auto pg = spg_t{pg_t{i, pool}};
+		coll_t cid(pg);
+
+		bool exists = os->collection_exists(cid);
+		auto ch = exists ?
+			os->open_collection(cid) :
+			os->create_new_collection(cid) ;
+
+		collections.emplace_back(pg, ch);
+
+		ObjectStore::Transaction t;
+		auto& coll = collections.back();
+		if (!exists) {
+			t.create_collection(coll.cid, split_bits);
+			ghobject_t pgmeta_oid(coll.pg.make_pgmeta_oid());
+			t.touch(coll.cid, pgmeta_oid);
+			int r = os->queue_transaction(coll.ch, std::move(t));
+			if (r) {
+				std::cerr << "Engine init failed with " << cpp_strerror(-r) << std::endl;
+				destroy_collections(os, collections);
+				return r;
+			}
+		}
+	}
+	return 0;
+}
+
+int destroy_collections(
+		std::unique_ptr<ObjectStore>& os,
+		std::vector<MyCollection>& collections)
+{
+	ObjectStore::Transaction t;
+	bool failed = false;
+	// remove our collections
+	for (auto& coll : collections) {
+		ghobject_t pgmeta_oid(coll.pg.make_pgmeta_oid());
+		t.remove(coll.cid, pgmeta_oid);
+		t.remove_collection(coll.cid);
+		int r = os->queue_transaction(coll.ch, std::move(t));
+		if (r && !failed) {
+			derr << "Engine cleanup failed with " << cpp_strerror(-r) << dendl;
+			failed = true;
+		}
+	}
+	return 0;
+}
+
+int create_object(std::unique_ptr<ObjectStore>& os, Object& o, size_t file_size)
+{
+	ObjectStore::Transaction t;
+
+	struct stat st;                                                           
+        int ret = os->stat(o.coll.ch, o.oid, &st);     
+	printf("stat: %d\n", ret);
+	t.touch(o.coll.cid, o.oid);
+	t.truncate(o.coll.cid, o.oid, file_size);
+
+	Result res;
+	t.register_on_applied(                                          
+                                new LambdaContext([&] (int r) {                 
+                                        res.signal(r);                           
+                                        })                                      
+                                );      
+	ret = os->queue_transaction(o.coll.ch, std::move(t));
+	res.wait();
+	return ret;
+}
+
+#define WAIT_COMMIT 0
+#define WAIT_APPLIED 1
+int write_object(std::unique_ptr<ObjectStore>& os, Object& o, off_t off, char *buf, size_t len, int wait)
+{
+	bufferlist bl;
+
+    	bl.push_back(buffer::copy(buf, len));
+	
+	ObjectStore::Transaction t;
+	Result r1;
+
+	t.write(o.coll.cid, o.oid, off, len, bl, 0);
+	if (wait == WAIT_COMMIT) {
+		t.register_on_commit(
+				new LambdaContext([&] (int r) {
+					r1.signal(r);
+					})
+				);
+	} else {
+		t.register_on_applied(
+				new LambdaContext([&] (int r) {
+					r1.signal(r);
+					})
+				);
+	}
+
+	int r = os->queue_transaction(o.coll.ch, std::move(t));
+	if (r) {
+		std::cerr << "can not queue transaction" << __func__ << endl;
+		return r;
+	}
+
+	r1.wait();
+	return r1.result();
+}
+
+int clone_range(std::unique_ptr<ObjectStore>& os,
+		Object& o1, Object& o2, uint64_t srcoff, uint64_t srclen,
+		uint64_t dstoff, int wait)
+{
+	ObjectStore::Transaction t;
+	Result r1;
+
+	t.clone_range(o1.coll.cid, o1.oid, o2.oid, srcoff, srclen, dstoff);
+	if (wait == WAIT_COMMIT) {
+		t.register_on_commit(
+			new LambdaContext([&] (int r) {
+				r1.signal(r);
+			})
+		);
+	} else {
+		t.register_on_applied(
+			new LambdaContext([&] (int r) {
+				r1.signal(r);
+			})
+		);
+	}
+
+	int r = os->queue_transaction(o1.coll.ch, std::move(t));
+	if (r) {
+		printf("%s: can not queue transaction\n", __func__);
+		return r;
+	}
+
+	r1.wait();
+	return r1.result();
+}
+	
+int benchmark_clone_range(std::unique_ptr<ObjectStore>& os, MyCollection &coll, int id, int repeat)
+{
+	char name[128];
+
+	snprintf(name, sizeof(name), "object_%d_0", id);
+	Object o1(name, coll);
+	snprintf(name, sizeof(name), "object_%d_1", id);
+	Object o2(name, coll);
+
+	char buf[64 * 1024];
+	memset(buf, 0, sizeof(buf));
+	strcpy(buf, "hello world");
+	int count = repeat;
+	int ret = 0;
+	auto start = ceph_clock_now();
+	while (count-- > 0) {
+		for (int i = 0; i < 128; ++i) {
+			ret = write_object(os, o1, i * sizeof(buf), buf, sizeof(buf), WAIT_COMMIT);
+			if (ret) {
+				printf("can not write object: %d\n", ret);
+				exit(1);
+			}
+
+			ret = write_object(os, o2, i * sizeof(buf), buf, sizeof(buf), WAIT_COMMIT);
+			if (ret) {
+				printf("can not write object: %d\n", ret);
+				exit(1);
+			}
+		}
+	}
+	auto end = ceph_clock_now();
+	auto elapsed1 = (double)(end - start);
+
+	count = repeat;
+	start = ceph_clock_now();
+	while (count-- > 0) {
+		for (int i = 0; i < 128; ++i) {
+			ret = write_object(os, o1, i * sizeof(buf), buf, sizeof(buf), WAIT_COMMIT);
+			if (ret) {
+				printf("can not write object: %d\n", ret);
+				exit(1);
+			}
+
+			ret = clone_range(os, o1, o2, i * sizeof(buf), sizeof(buf), i * sizeof(buf), WAIT_COMMIT);
+			if (ret) {
+				printf("can not clone_range: %d\n", ret);
+				exit(1);
+			}
+		}
+	}
+	end = ceph_clock_now();
+	auto elapsed2 = (double)(end - start);
+	printf("thread 1: double write: %f, clone_range: %f\n", elapsed1, elapsed2);
+	return 0;
+}
+
+// test clone_range correctness
+int test_clone_range(std::unique_ptr<ObjectStore>& os, MyCollection &coll)
+{
+	Object o1("object1", coll);
+	Object o2("object2", coll);
+
+	int ret = create_object(os, o1, 0);
+	if (ret) {
+		printf("can not create object1\n");
+		exit(1);
+	}
+	ret = create_object(os, o2, 0);
+	if (ret) {
+		printf("can not create object2\n");
+		exit(1);
+	}
+
+	char buf[64 * 1024];
+	char buf2[64 * 1024]; // same size as buf
+	memset(buf, 0, sizeof(buf));
+	strcpy(buf, "hello world");
+
+	ret = write_object(os, o1, 512, buf, sizeof(buf), WAIT_COMMIT);
+	if (ret) {
+		printf("can not write object: %d\n", ret);
+		exit(1);
+	}
+	bufferlist bl;
+    	ret = os->read(o1.coll.ch, o1.oid, 512, sizeof(buf), bl);
+	if (ret != sizeof(buf)) {
+		printf("can not read object: %d\n", ret);
+		exit(1);
+	}
+	if (!bl.contents_equal(buf, sizeof(buf))) {
+		printf("inconst data with buf\n");
+		exit(1);
+	}
+
+	ret = clone_range(os, o1, o2, 512, sizeof(buf), 0, WAIT_COMMIT);
+	if (ret) {
+		printf("clone range error, %d\n", ret);
+		exit(1);
+	}
+
+	memcpy(buf2, buf, sizeof(buf));
+
+	memset(buf, 0, sizeof(buf));
+	strcpy(buf, "xxxx");
+
+	// overwrite object1 
+	ret = write_object(os, o1, 512, buf, sizeof(buf), WAIT_COMMIT);
+	if (ret) {
+		printf("can not write object: %d\n", ret);
+		exit(1);
+	}
+
+	bl.clear();
+
+	// read object2
+    	ret = os->read(o1.coll.ch, o2.oid, 0, sizeof(buf), bl);
+	if (ret != sizeof(buf)) {
+		printf("can not read object: %d\n", ret);
+		exit(1);
+	}
+
+	// check if o2 is changed, should not!
+	if (!bl.contents_equal(buf2, sizeof(buf2))) {
+		printf("inconsistent data with buf2 after ovewritting o1\n");
+		exit(1);
+	}
+	printf("test clone_range success\n");
+	return 0;
 }
 
 int main(int argc, const char **argv)
@@ -67,11 +432,30 @@ int main(int argc, const char **argv)
 			   "bluestore",
                            g_conf().get_val<std::string>("osd data"),
                            g_conf().get_val<std::string>("osd journal"));
-	printf("%p\n", os.get());
+	if (!os) {
+		std::cerr << " can not open object store" << std::endl;
+		exit(1);
+	}
+
 	int ret = os->mkfs();
 	printf("mkfs ret=%d\n", ret);
+	if (ret) {
+		exit(1);
+	}
 	ret = os->mount();
-	printf("mount ret=%d\n", ret);
+
+	init_collections(os, MyCollection::MIN_POOL_ID,
+			 collections, COLLECTION_COUNT);
+
+	test_clone_range(os, collections[0]);
+
+	benchmark_clone_range(os, collections[0], 0, 100);
+
+
+	// destructor MyCollection memory obj before umount,
+	// otherwise there is illegal reference to object store
+	collections.clear();
+
 	ret = os->umount();
 	printf("umount ret=%d\n", ret);
 }
